+++ TMessagesProj/src/main/java/org/telegram/messenger/exoplayer2/source/dash/manifest/SegmentBase.java
@@ -99,7 +99,7 @@ public abstract class SegmentBase {
    */
   public abstract static class MultiSegmentBase extends SegmentBase {
 
-    /* package */ final int startNumber;
+    /* package */ final long startNumber;
     /* package */ final long duration;
     /* package */ final List<SegmentTimelineElement> segmentTimeline;
 
@@ -111,43 +111,46 @@ public abstract class SegmentBase {
      *     division of this value and {@code timescale}.
      * @param startNumber The sequence number of the first segment.
      * @param duration The duration of each segment in the case of fixed duration segments. The
-     *     value in seconds is the division of this value and {@code timescale}. If
-     *     {@code segmentTimeline} is non-null then this parameter is ignored.
+     *     value in seconds is the division of this value and {@code timescale}. If {@code
+     *     segmentTimeline} is non-null then this parameter is ignored.
      * @param segmentTimeline A segment timeline corresponding to the segments. If null, then
      *     segments are assumed to be of fixed duration as specified by the {@code duration}
      *     parameter.
      */
-    public MultiSegmentBase(RangedUri initialization, long timescale, long presentationTimeOffset,
-        int startNumber, long duration, List<SegmentTimelineElement> segmentTimeline) {
+    public MultiSegmentBase(
+        RangedUri initialization,
+        long timescale,
+        long presentationTimeOffset,
+        long startNumber,
+        long duration,
+        List<SegmentTimelineElement> segmentTimeline) {
       super(initialization, timescale, presentationTimeOffset);
       this.startNumber = startNumber;
       this.duration = duration;
       this.segmentTimeline = segmentTimeline;
     }
 
-    /**
-     * @see DashSegmentIndex#getSegmentNum(long, long)
-     */
-    public int getSegmentNum(long timeUs, long periodDurationUs) {
-      final int firstSegmentNum = getFirstSegmentNum();
-      final int segmentCount = getSegmentCount(periodDurationUs);
+    /** @see DashSegmentIndex#getSegmentNum(long, long) */
+    public long getSegmentNum(long timeUs, long periodDurationUs) {
+      final long firstSegmentNum = getFirstSegmentNum();
+      final long segmentCount = getSegmentCount(periodDurationUs);
       if (segmentCount == 0) {
         return firstSegmentNum;
       }
       if (segmentTimeline == null) {
         // All segments are of equal duration (with the possible exception of the last one).
         long durationUs = (duration * C.MICROS_PER_SECOND) / timescale;
-        int segmentNum = startNumber + (int) (timeUs / durationUs);
+        long segmentNum = startNumber + timeUs / durationUs;
         // Ensure we stay within bounds.
         return segmentNum < firstSegmentNum ? firstSegmentNum
             : segmentCount == DashSegmentIndex.INDEX_UNBOUNDED ? segmentNum
             : Math.min(segmentNum, firstSegmentNum + segmentCount - 1);
       } else {
         // The index cannot be unbounded. Identify the segment using binary search.
-        int lowIndex = firstSegmentNum;
-        int highIndex = firstSegmentNum + segmentCount - 1;
+        long lowIndex = firstSegmentNum;
+        long highIndex = firstSegmentNum + segmentCount - 1;
         while (lowIndex <= highIndex) {
-          int midIndex = lowIndex + (highIndex - lowIndex) / 2;
+          long midIndex = lowIndex + (highIndex - lowIndex) / 2;
           long midTimeUs = getSegmentTimeUs(midIndex);
           if (midTimeUs < timeUs) {
             lowIndex = midIndex + 1;
@@ -161,12 +164,10 @@ public abstract class SegmentBase {
       }
     }
 
-    /**
-     * @see DashSegmentIndex#getDurationUs(int, long)
-     */
-    public final long getSegmentDurationUs(int sequenceNumber, long periodDurationUs) {
+    /** @see DashSegmentIndex#getDurationUs(long, long) */
+    public final long getSegmentDurationUs(long sequenceNumber, long periodDurationUs) {
       if (segmentTimeline != null) {
-        long duration = segmentTimeline.get(sequenceNumber - startNumber).duration;
+        long duration = segmentTimeline.get((int) (sequenceNumber - startNumber)).duration;
         return (duration * C.MICROS_PER_SECOND) / timescale;
       } else {
         int segmentCount = getSegmentCount(periodDurationUs);
@@ -177,14 +178,13 @@ public abstract class SegmentBase {
       }
     }
 
-    /**
-     * @see DashSegmentIndex#getTimeUs(int)
-     */
-    public final long getSegmentTimeUs(int sequenceNumber) {
+    /** @see DashSegmentIndex#getTimeUs(long) */
+    public final long getSegmentTimeUs(long sequenceNumber) {
       long unscaledSegmentTime;
       if (segmentTimeline != null) {
-        unscaledSegmentTime = segmentTimeline.get(sequenceNumber - startNumber).startTime
-            - presentationTimeOffset;
+        unscaledSegmentTime =
+            segmentTimeline.get((int) (sequenceNumber - startNumber)).startTime
+                - presentationTimeOffset;
       } else {
         unscaledSegmentTime = (sequenceNumber - startNumber) * duration;
       }
@@ -195,14 +195,12 @@ public abstract class SegmentBase {
      * Returns a {@link RangedUri} defining the location of a segment for the given index in the
      * given representation.
      *
-     * @see DashSegmentIndex#getSegmentUrl(int)
+     * @see DashSegmentIndex#getSegmentUrl(long)
      */
-    public abstract RangedUri getSegmentUrl(Representation representation, int index);
+    public abstract RangedUri getSegmentUrl(Representation representation, long index);
 
-    /**
-     * @see DashSegmentIndex#getFirstSegmentNum()
-     */
-    public int getFirstSegmentNum() {
+    /** @see DashSegmentIndex#getFirstSegmentNum() */
+    public long getFirstSegmentNum() {
       return startNumber;
     }
 
@@ -235,15 +233,20 @@ public abstract class SegmentBase {
      *     division of this value and {@code timescale}.
      * @param startNumber The sequence number of the first segment.
      * @param duration The duration of each segment in the case of fixed duration segments. The
-     *     value in seconds is the division of this value and {@code timescale}. If
-     *     {@code segmentTimeline} is non-null then this parameter is ignored.
+     *     value in seconds is the division of this value and {@code timescale}. If {@code
+     *     segmentTimeline} is non-null then this parameter is ignored.
      * @param segmentTimeline A segment timeline corresponding to the segments. If null, then
      *     segments are assumed to be of fixed duration as specified by the {@code duration}
      *     parameter.
      * @param mediaSegments A list of {@link RangedUri}s indicating the locations of the segments.
      */
-    public SegmentList(RangedUri initialization, long timescale, long presentationTimeOffset,
-        int startNumber, long duration, List<SegmentTimelineElement> segmentTimeline,
+    public SegmentList(
+        RangedUri initialization,
+        long timescale,
+        long presentationTimeOffset,
+        long startNumber,
+        long duration,
+        List<SegmentTimelineElement> segmentTimeline,
         List<RangedUri> mediaSegments) {
       super(initialization, timescale, presentationTimeOffset, startNumber, duration,
           segmentTimeline);
@@ -251,8 +254,8 @@ public abstract class SegmentBase {
     }
 
     @Override
-    public RangedUri getSegmentUrl(Representation representation, int sequenceNumber) {
-      return mediaSegments.get(sequenceNumber - startNumber);
+    public RangedUri getSegmentUrl(Representation representation, long sequenceNumber) {
+      return mediaSegments.get((int) (sequenceNumber - startNumber));
     }
 
     @Override
@@ -284,8 +287,8 @@ public abstract class SegmentBase {
      *     division of this value and {@code timescale}.
      * @param startNumber The sequence number of the first segment.
      * @param duration The duration of each segment in the case of fixed duration segments. The
-     *     value in seconds is the division of this value and {@code timescale}. If
-     *     {@code segmentTimeline} is non-null then this parameter is ignored.
+     *     value in seconds is the division of this value and {@code timescale}. If {@code
+     *     segmentTimeline} is non-null then this parameter is ignored.
      * @param segmentTimeline A segment timeline corresponding to the segments. If null, then
      *     segments are assumed to be of fixed duration as specified by the {@code duration}
      *     parameter.
@@ -294,9 +297,15 @@ public abstract class SegmentBase {
      *     null then {@code initialization} will be used.
      * @param mediaTemplate A template defining the location of each media segment.
      */
-    public SegmentTemplate(RangedUri initialization, long timescale, long presentationTimeOffset,
-        int startNumber, long duration, List<SegmentTimelineElement> segmentTimeline,
-        UrlTemplate initializationTemplate, UrlTemplate mediaTemplate) {
+    public SegmentTemplate(
+        RangedUri initialization,
+        long timescale,
+        long presentationTimeOffset,
+        long startNumber,
+        long duration,
+        List<SegmentTimelineElement> segmentTimeline,
+        UrlTemplate initializationTemplate,
+        UrlTemplate mediaTemplate) {
       super(initialization, timescale, presentationTimeOffset, startNumber,
           duration, segmentTimeline);
       this.initializationTemplate = initializationTemplate;
@@ -315,10 +324,10 @@ public abstract class SegmentBase {
     }
 
     @Override
-    public RangedUri getSegmentUrl(Representation representation, int sequenceNumber) {
+    public RangedUri getSegmentUrl(Representation representation, long sequenceNumber) {
       long time;
       if (segmentTimeline != null) {
-        time = segmentTimeline.get(sequenceNumber - startNumber).startTime;
+        time = segmentTimeline.get((int) (sequenceNumber - startNumber)).startTime;
       } else {
         time = (sequenceNumber - startNumber) * duration;
       }
